requirements.txt
Flask
google-cloud-aiplatform
google-cloud-firestore
google-auth
google-auth-oauthlib
gunicorn

